feat: add institutions index route and refactor institution details route
